handle post when more attributes are posted

+++ turbobee_app/views.py
@@ -53,14 +53,17 @@ def store(bibcode):
             msg = TurboBeeMsg.loads('adsmsg.turbobee.TurboBeeMsg', req_file)
 
             page_d = {}
-            page_d['qid'] = hashlib.sha256(str(msg)).hexdigest()
+            page_d['content'] = str(msg.get_value())
+            page_d['qid'] = hashlib.sha256(page_d['content']).hexdigest()
             ts = msg.get_timestamp()
             page_d['created'] = dt.datetime.fromtimestamp(ts.seconds + ts.nanos * 10**-9) 
-            page_d['content'] = msg.get_value()
             page_d['content_type'] = 'application/' + ctypes[msg.ctype]
-            page_d['updated'] = page_d['created']
-            page_d['expires'] = page_d['created'] + dt.timedelta(days=365)
-            page_d['lifetime'] = page_d['created'] + dt.timedelta(days=365*100)
+            page_d['updated'] = msg.updated.ToDatetime() if msg.updated.ToSeconds() != 0 \
+                else page_d['created']
+            page_d['expires'] = msg.expires.ToDatetime() if msg.expires.ToSeconds() != 0 \
+                else page_d['created'] + dt.timedelta(days=365)
+            page_d['lifetime'] = msg.eol.ToDatetime() if msg.eol.ToSeconds() != 0 \
+                else page_d['created'] + dt.timedelta(days=365*100)
 
             page = Pages(**page_d)
 
@@ -74,12 +77,12 @@ def store(bibcode):
 
             return str(msg), 200
         elif request.method == 'DELETE':
-            try:
-                pages = session.query(Pages).filter_by(qid=bibcode)
-                pages.first().toJSON() # test whether any page was found
-                pages.delete()
-            except:
+            pages = session.query(Pages).filter_by(qid=bibcode)
+            if len(pages.all()) == 0:
                 return abort(404)
+            else:
+                pages.delete()
+
             try:
                 session.commit()
             except exc.IntegrityError as e:
